+++ fhir-mapbuilder-validation/src/main/java/fr/aphp/mapbuilder/model/ParsingError.java
@@ -1,0 +1,7 @@
+package fr.aphp.mapbuilder.model;
+
+public class ParsingError extends Exception {
+    public ParsingError(String message, Throwable cause) {
+        super(message, cause);
+    }
+}