+++ vscode-extension/src/extension.ts
@@ -5,20 +5,9 @@ import {FhirDefinition} from './FhirDefinition';
 import {FmlValidation} from './FmlValidation';
 import {MapBuilderJavaProcess} from "./MapBuilderJavaProcess";
 import {MapBuilderValidationApi} from "./MapBuilderValidationApi";
+import {MapBuilderWatcher} from "./MapBuilderWatcher";
 
 const FML_MODE = {language: 'fml', scheme: 'file'};
-
-function getMapBuilderValidationApi(validationOutputChannel: OutputChannel) {
-    const mapBuilderJavaProcess = new MapBuilderJavaProcess(validationOutputChannel);
-
-    const api = new MapBuilderValidationApi(validationOutputChannel);
-    api.isAppRunning().then(isAppRunning => {
-        if (!isAppRunning) {
-            mapBuilderJavaProcess.start();
-        }
-    });
-    return api;
-}
 
 export function activate(context: vscode.ExtensionContext): {
     completionProviderInstance: FmlCompletionProvider | null
@@ -40,6 +29,8 @@ export function activate(context: vscode.ExtensionContext): {
 
         addValidationAfterLoadingPackageCommand(principalChannel, detailsChannel, api, context);
 
+        addWatcher(principalChannel, api);
+
         return {completionProviderInstance};
 
     } catch (err) {
@@ -54,7 +45,25 @@ export function deactivate() {
 
 }
 
-export function addAutoComplete(outputChannel: OutputChannel, context: vscode.ExtensionContext): [FhirDefinition, FmlCompletionProvider] {
+function getMapBuilderValidationApi(validationOutputChannel: OutputChannel) {
+    const mapBuilderJavaProcess = new MapBuilderJavaProcess(validationOutputChannel);
+
+    const api = new MapBuilderValidationApi(validationOutputChannel);
+    api.isAppRunning().then(isAppRunning => {
+        if (!isAppRunning) {
+            mapBuilderJavaProcess.start();
+        }
+    });
+    return api;
+}
+
+function addWatcher(principalChannel: OutputChannel, api: MapBuilderValidationApi) {
+    const watcher = new MapBuilderWatcher(principalChannel, api);
+    watcher.watchFmlFiles();
+    watcher.watchIgPackage();
+}
+
+function addAutoComplete(outputChannel: OutputChannel, context: vscode.ExtensionContext): [FhirDefinition, FmlCompletionProvider] {
     const fhirDefinitionInstance = new FhirDefinition(outputChannel);
     const completionProviderInstance = new FmlCompletionProvider(fhirDefinitionInstance, outputChannel);
 
@@ -65,7 +74,7 @@ export function addAutoComplete(outputChannel: OutputChannel, context: vscode.Ex
     return [fhirDefinitionInstance, completionProviderInstance];
 }
 
-export function addFMLTemplate(context: vscode.ExtensionContext) {
+function addFMLTemplate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('mapbuilder.InsertTemplate', () => {
         if (vscode.window.activeTextEditor) {
             vscode.languages.setTextDocumentLanguage(vscode.window.activeTextEditor.document, FML_MODE.language as string);
@@ -74,7 +83,7 @@ export function addFMLTemplate(context: vscode.ExtensionContext) {
     }));
 }
 
-export function addValidationCommand(outputChannel: OutputChannel, validationOutputChannel: OutputChannel, mapBuilderValidationApi: MapBuilderValidationApi, context: vscode.ExtensionContext) {
+function addValidationCommand(outputChannel: OutputChannel, validationOutputChannel: OutputChannel, mapBuilderValidationApi: MapBuilderValidationApi, context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('mapbuilder.Validation', async () => {
 
         const fmlValidation = new FmlValidation(outputChannel, mapBuilderValidationApi);
@@ -82,7 +91,7 @@ export function addValidationCommand(outputChannel: OutputChannel, validationOut
     }));
 }
 
-export function addValidationWithDefaultFilesCommand(outputChannel: OutputChannel, validationOutputChannel: OutputChannel, mapBuilderValidationApi: MapBuilderValidationApi, context: vscode.ExtensionContext) {
+function addValidationWithDefaultFilesCommand(outputChannel: OutputChannel, validationOutputChannel: OutputChannel, mapBuilderValidationApi: MapBuilderValidationApi, context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('mapbuilder.ValidationWithDefaultFiles', async () => {
 
         const fmlValidation = new FmlValidation(outputChannel, mapBuilderValidationApi);
@@ -91,7 +100,7 @@ export function addValidationWithDefaultFilesCommand(outputChannel: OutputChanne
     }));
 }
 
-export function addValidationAfterLoadingPackageCommand(
+function addValidationAfterLoadingPackageCommand(
     outputChannel: OutputChannel,
     validationOutputChannel: OutputChannel,
     mapBuilderValidationApi: MapBuilderValidationApi,