+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-# 1.0.0 (2024-02-05)
+# 1.0.0 (2025-04-07)
 
-* Refactoring of the Matchbox application: called fhir-mapbuilder-validation via REST requests and optimized.
+* Parse .fml files and/or load the IG package when an add or change event occurs.
+* Refactoring: call fhir-mapbuilder-validation via REST requests and optimized.
 * Added a new command to validate the StructureMap and edit test parameters.
 * Added a command to reset and load the Matchbox engine and load the current IG package.